// web-app/src/app/core/models/avance.ts
export class Avance {
  id?: number;
  membreId?: number;
  duree?: number;
  montantCapital?: number;
  dateDemande?: string;
  dateDecaissement?: string;

  montantCapitalRestant?: number;
  status?: string;
  nombreEcheancePaye?: number;
  nombreEcheanceImpaye?: number;
  dateDerniereEcheance?: string;

  nom?: string;
  photo?: string;
  nomSexe?: string;
  agenceId?: number;
}